requirements.txt
pydub
google-cloud-texttospeech
python-dotenv
openai
gTTS
